add repeatcount and helper thread

+++ libsoundpoolcompat/src/main/jni/AudioSource.h
@@ -44,8 +44,8 @@ namespace SoundPoolCompat {
 
 
         int _fd;
-        int _fd_offset;
-        int _fd_length;
+        int64_t _fd_offset;
+        int64_t _fd_length;
         bool _fd_autoclose;
 
         std::string _uri_path;
@@ -62,7 +62,7 @@ namespace SoundPoolCompat {
         static bool setAudioSourcePCM(int audioID,int numChannels,int samplingRate,int bitPerSample);
         static bool addPCMBuffer(int audioID,void* pBuf,int offset,int size);
 
-        static bool setAudioSourceFileDescriptor(int audioID,int fd,int offset,int length,bool autoclose = true);
+        static bool setAudioSourceFileDescriptor(int audioID,int fd,int64_t offset,int64_t length,bool autoclose = true);
         static bool setAudioSourceURI(int audioID,const std::string& uri);
 
     };
